requirements.txt
python-docx>=1.1.0
PyYAML>=6.0.1
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.3
python-multipart>=0.0.9
